Move code to lib/jsonschema-tools.js; use index.js for explicit module exports
